refactor: move AccessorMixin to accessor module
